third_party/WebKit/LayoutTests/http/tests/devtools/console/console-last-result-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that console exposes last evaluation result as $_.


Running: testLastResult
1+1
2
$_
2

Running: testLastResultAfterConsoleClear
$_
undefined
